+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * update typescript definition to new options
 * fix cursor for PrismJS punctuation class
 * fix emoji that contain U+FE0F character at the end [#424](https://github.com/jcubic/jquery.terminal/issues/424)
+* fix for combine characters
 
 ## 1.21.0
 ### Security